Completed Search Asset Component
